# docs/openapi.yaml
openapi: 3.0.3
info:
  title: vocagame technical testing backend
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/product/create:
    post:
      summary: create new product
      description: add new product to database
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
            examples:
              all parameters:
                description: input all parameter
                value:
                  name: lamp
                  description: led lamp from philips with 3 years of warranty
                  price: 5
                  availability: 27
              only name:
                description: input all parameter
                value:
                  name: laptop
      responses:
        "201":
          description: created product data with id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseProduct"
              examples:
                all parameter:
                  description: fill all parameter in request body
                  value:
                    description: create new product success
                    error_message: ''
                    execute_at: 2024/11/04 15:59:03.626
                    result: success
                    value:
                      id: 94
                      name: lamp
                      description: have 3 years warranty
                      price: 5
                      availability: 10
                only name:
                  description: fill all parameter in request body
                  value:
                    description: create new product success
                    error_message: ''
                    execute_at: 2024/11/04 15:59:03.626
                    result: success
                    value:
                      id: 97
                      name: laptop
                      description: ""
                      price: 0
                      availability: 0
        "422":
          description: Unprocessably Entity, when input is invalid
          content:
            application/json:
              schema:
                $ref: "#components/schemas/error"
              examples:
                empty name:
                  description: name is empty in request body
                  value:
                    description: 
                    - Name is a required field
                    error_message: Unauthorized
                    execute_at: 2024/11/04 20:50:08.938
                    result: failure
                invalid input:
                  description: input is not following requirement
                  value:
                    description: 
                    - Price must be 0 or greater
                    - Availability must be 0 or greater
                    error_message: Unauthorized
                    execute_at: 2024/11/04 20:50:08.938
                    result: failure
        "409":
          description: Conflict, using a name that already exists in database
          content:
            application/json:
              schema:
                $ref: "#components/schemas/error"
              example:
                description: 
                - duplicate data
                error_message: Conflict
                execute_at: 2024/11/04 20:50:08.938
                result: failure
        "401":
          description: Unauthorized, make request with expired access token
          content:
            application/json:
              schema:
                $ref: "#components/schemas/error"
              example:
                description: 
                - 'failed to parse token when verifying, err: token has invalid claims: token is expired'
                error_message: Conflict
                execute_at: 2024/11/04 20:50:08.938
                result: Unauthorized
  /api/v1/product/get/{id}:
    get:
      summary: get product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The product ID
      responses:
        "200":
          description: OK, success get product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseProduct"
              example:
                description: success get product
                error_message: ''
                execute_at: 2024/11/04 15:59:03.626
                result: success
                value:
                  id: 94
                  name: lamp
                  description: have 3 years warranty
                  price: 5
                  availability: 10
        "400":
          description: Bad Request, no product found in database
          content:
            application/json:
              schema:
                $ref: "#components/schemas/error"
              example:
                description: 
                - no rows in result set
                error_message: Bad Request
                execute_at: 2024/11/04 20:50:08.938
                result: failure
        "401":
          description: Unauthorized, request without auth header
          content:
            application/json:
              schema:
                $ref: "#components/schemas/error"
              example:
                description: 
                - no authorization header found
                error_message: Unauthorized
                execute_at: 2024/11/04 20:50:08.938
                result: failure
  /api/v1/product/list?{page}&{limit}:
    get:
      summary: list all products
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
          example: 3
          description: data for current page
        - in: query
          name: limit
          required: false
          schema:
            type: integer
          example: 3
          description: the number of products one to display per page
      responses:
        200:
          description: list of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseListProduct'
              examples:
                Data:
                  description: success with data
                  value:
                    description: "list of books"
                    error_message: ""
                    execute_at: "2024/11/05 01:58:18.341"
                    pagination: 
                      current_page: "3"
                      total_pages: "4"
                    result: "success"
                    value: 
                    - id: "196"
                      name: "ihxkcbs"
                      description: "ttbhoubaiokiszzbrhfyfutgezugfqnrwxpcmqcwqsstgkgewj"
                      price: "467"
                      availability: "28"
                    - id: "197"
                      name: "myngees"
                      description: "isruskqxyqgzfituknqiyhldptzgzpdjnlpurikqhagmrwutqc"
                      price: "100"
                      availability: "36"
                    - id: "198"
                      name: "wgckpte"
                      description: "favziqlioilnfpwthlpxkjhozxlxsjneygvtkaorpapufpirbg"
                      price: "455"
                      availability: "9"
                No Data:
                  description: success with data
                  value:
                    description: "list of books"
                    error_message: ""
                    execute_at: "2024/11/05 01:58:18.341"
                    pagination: 
                      current_page: "3"
                      total_pages: "1"
                    result: "success"
                    value: 
                Default Params:
                  description: 'using default params if params not filled or negative, default: page = 1 & limit = 10'
                  value:
                    description: "list of books"
                    error_message: ""
                    execute_at: "2024/11/05 01:58:45.612"
                    pagination: 
                      current_page: "1"
                      total_pages: "1"
                    result: "success"
                    value: 
                    - id: "190"
                      name: "bxrfjmd"
                      description: "zoidmoxtqsncnlmqogkqmpvyraceswhkvfduatdfdjaprolqza"
                      price: "398"
                      availability: "3"
                    - id: "191"
                      name: "hzpelpg"
                      description: "ewelaruhanevcuvixfufnorhdfbqkcgahsvlzycezymbpfyvzn"
                      price: "268"
                      availability: "43"
                    - id: "192"
                      name: "pcrqbya"
                      description: "sglmfmnxobfjghkwetwfxbepznvwppmvdcgelseanhxbzznrmu"
                      price: "486"
                      availability: "32"
                    - id: "193"
                      name: "bucnxtw"
                      description: "nixnpfayufhfstajrcjrxnvslydargsyjtclklefvmembemeln"
                      price: "350"
                      availability: "17"
                    - id: "194"
                      name: "bmitnqt"
                      description: "eglqlxhuoucpaytshtwtaezxjhqzzefzbqtyvlrdimnszbedqe"
                      price: "190"
                      availability: "2"
                    - id: "195"
                      name: "ilppmyg"
                      description: "miyshlfcnoapngtparyldanzdibavzdfwbrrwpgnleatwuxmlx"
                      price: "496"
                      availability: "2"
                    - id: "196"
                      name: "ihxkcbs"
                      description: "ttbhoubaiokiszzbrhfyfutgezugfqnrwxpcmqcwqsstgkgewj"
                      price: "467"
                      availability: "28"
                    - id: "197"
                      name: "myngees"
                      description: "isruskqxyqgzfituknqiyhldptzgzpdjnlpurikqhagmrwutqc"
                      price: "100"
                      availability: "36"
                    - id: "198"
                      name: "wgckpte"
                      description: "favziqlioilnfpwthlpxkjhozxlxsjneygvtkaorpapufpirbg"
                      price: "455"
                      availability: "9"
                    - id: "199"
                      name: "xgrjvdn"
                      description: "jpjnrznttfkvfvmhiffugvnimxgjpmqvynloooyuzdcavnxqmz"
                      price: "92"
                      availability: "28"
...
  /api/v1/product/delete/{id}:
    delete:
      summary: delete product
      description: delete a product using product ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
          example: 95
          description: The product ID
      responses:
        "200":
          description: OK, success get product
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      value:
                        type: string
              example:
                description: success deleted product
                error_message: ''
                execute_at: 2024/11/04 15:59:03.626
                result: success
                value: ""
        "400":
          description: Bad Request, no product found in database
          content:
            application/json:
              schema:
                $ref: "#components/schemas/error"
              example:
                description: 
                - 'failed to delete product, err: no row deleted'
                error_message: Bad Request
                execute_at: 2024/11/04 20:50:08.938
                result: failure
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaseResponse:
      type: object
      properties:
        description:
          type: string
        error_message:
          type: string
        execute_at:
          type: string
        result:
          type: string
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        availability:
          type: integer
      required:
        - id
    pagination:
      type: object
      properties:
        current_page:
          type: integer
        total_page:
          type: integer
    ProductRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          minLength: 0
        price:
          type: integer
          minimum: 0
        availability:
          type: integer
          minimum: 0
      required:
        - name
...
    ResponseProduct:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            value:
              $ref: '#/components/schemas/Product'
    ResponseListProduct:
      allOf:
        - $ref: '#components/schemas/BaseResponse'
        - type: object
          properties:
            pagination:
              $ref: '#/components/schemas/pagination'
        - type: array
          properties:
            value:
              $ref: '#/components/schemas/Product'